# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticates a user using provided credentials.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Handle user login",
                "parameters": [
                    {
                        "description": "User credentials for login",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.Credential"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/example.SuccessfulLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/example.InvalidCredentialResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "example.InvalidCredentialResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "bad_request"
                        },
                        "message": {
                            "type": "string",
                            "example": "invalid credential"
                        }
                    }
                }
            }
        },
        "example.SuccessfulLoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "access_token": {
                            "type": "string",
                            "example": "your_access_token"
                        },
                        "refresh_token": {
                            "type": "string",
                            "example": "your_refresh_token"
                        }
                    }
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string",
                            "example": "ok"
                        },
                        "message": {
                            "type": "string",
                            "example": "Ok"
                        }
                    }
                }
            }
        },
...
        "user.Credential": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@klondike.id"
                },
                "password": {
                    "type": "string",
                    "example": "1234qweR!"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}